DO NOT MERGE TO MASTER: finish
This is a super quick proof-of-concept that I want more eyeballs looking
at before I put it on master.  My plan is to let a bunch of the regular
meeting attendees know about this branch and to get it cleaned up in the
next couple of days and then get it merged into master and then to use the
Meetup mailing list to let everyone else know.