Includes the first draft of my Talking to One's Self presentation.
I haven't figured out the best way to package inputs for markdeck
separate from markdeck itself.  Right now I'm just using a .gitignore
for everything other than slides.md and my sample .rb files, because
it seems weird to commit all the markdeck implementation files with
the presentation itself.

There's potential for a big downside if I (accidentally?) edit any of
the files that are in .gitignore, because that could make it so that
the presentation is not reproducible from source.  OTOH, it seems
unlikely that I'll do so and perhaps once I've written my presentation
I'll gain enlightenment on markdeck structure.

FWIW, I tried making slides.md a symlink, but that didn't work.